fix: add JPA implementation for SubjectPersistenceAdapter
Problem:
The application didn't start because spring could
not autowire the beans for JPA components. JPA
components was missing.

Solution:
Create JPA implementation the subject persistence adpater.
- Implement the SubjectJpaAdapter interface
- Create a Subject JPA entity
- Create the Subject repository
- Create a mapper to convert Core Entity to JPA Entity

In order to keep the core application decoupled, it was
created JPA entities which will be used in JPA operations.
A mapper was criated to convert Core Entity into
JPA Entity and vice versa.

Obs:
Still need to create the remaining JPA componentes.