fix: remove AssemblyApplicationTests that causes test to fail
Problem:
Test was failing because AssemblyApplicationTests was
annotated with @SpringBootTest wich tells Spring Boot
to load the whole application context and it ended up
failing because Spring tried to connect to database
wich should not be used for unit and REST
integration tests.

Solution:
The class AssemblyApplicationTests was removed
because it is empty. A better solution is to
set maven profiles, but it will be done later on.